speed up type checking
